Fixed extractors not being found on some systems
On some systems many or even all helper binaries could not be found and therefore extraction failed completely. This is most likely due to the use of relative paths to the executables. Whenever UniExtract was started the whole /bin directory was temporarily added to PATH using EnvSet. Instead, absolute paths are now used when a file is executed. This also fixes a couple of other issues from previous versions and improves code quality.

  Fixed cleanup for .chm and EnigmaVB files
  Fixed error message after failed iso extraction: choices were mixed-up
  Fixed missing file type information in NOTPACKED error message for FFMPEG extraction
  Fixed missing packer name in unpack prompt
  Changed .arj extraction to use 7z instead of arj
  Changed .hlp extraction: improved logging, localized file name
  Removed redundant 64-bit MediaInfo dll