Fixed crash when trying to redownload currupt files
If no language files can be found by UniExtract, it asks to repair the installation by downloading any missing files. This commit fixes a hard crash right after the prompt to redownload.

RC 2 has been re-released with this issue fixed. An update is available for all existing installations of UniExtract.